Fixing integer overflow issue in RTS metrics (#6539)

+++ src/WebJobs.Script.WebHost/Scale/TableEntityConverter.cs
@@ -91,7 +91,9 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Scale
                     entityProperty = new EntityProperty(value.ToObject<Guid>());
                     return true;
                 case JTokenType.Integer:
-                    entityProperty = new EntityProperty(value.ToObject<int>());
+                    // to handle both ints and longs, we normalize integer values
+                    // to type long
+                    entityProperty = new EntityProperty(value.ToObject<long>());
                     return true;
                 case JTokenType.String:
                     entityProperty = new EntityProperty(value.ToObject<string>());
